% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addCumPos}
\alias{addCumPos}
\title{Add cumulative genomic position}
\usage{
addCumPos(dat, ref, dat.type)
}
\arguments{
\item{dat}{Input data, either a seg dataframe or data}

\item{ref}{GRanges object from getChrLength}

\item{dat.type}{either 'data' or 'seg'}
}
\description{
Add cumulative genomic position
}
