% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alphaIt}
\alias{alphaIt}
\title{Add an apha layer to the matrix}
\usage{
alphaIt(matX, col, alpha = 0.5)
}
\arguments{
\item{matX}{A matrix of RGB colours ranging in value from 0 to 255}

\item{col}{Colour to add to the matrix}

\item{alpha}{Alpha value of the coloure}
}
\description{
Add an apha layer to the matrix
}
