% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangesHelper.R
\name{calcCumulativeGPos}
\alias{calcCumulativeGPos}
\title{Calculate cumulative genomic positions}
\usage{
calcCumulativeGPos(gr0, seqstyle = "ENSEMBL")
}
\arguments{
\item{gr0}{GenomicRanges object to get the cumulative positions for}

\item{seqstyle}{seq style from GenomeInfoDb (Default='ENSEMBL')}
}
\value{
GRanges object with the cumulative postiions added to it
}
\description{
Calculate cumulative genomic positions
}
