% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFC.R
\name{genHC}
\alias{genHC}
\title{Initiailize HilbertCurve object}
\usage{
genHC(end, order)
}
\arguments{
\item{end}{Typically, the size of the genome [integer]}

\item{order}{Order to generate hilbert curve for [integer]}
}
\value{
an object from HilbertCurve::HilbertCurve()
}
\description{
Wrapper function to run the HilbertCurve() function
using only a set end point and order
}
