% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChrLength}
\alias{getChrLength}
\title{hg19 chromosome sizes}
\usage{
getChrLength()
}
\value{
GenomicRanges object for the hg19 chromosomes and cumulative start/end points
}
\description{
Returns a GRanges object of the hg19 chromosome sizes
as well as their cumulative starts, ends, and midpoints
}
