% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangesHelper.R
\name{intersectAndAnnotate}
\alias{intersectAndAnnotate}
\title{Intersect and annotate sample with reference GR}
\usage{
intersectAndAnnotate(s_gr, ref_gr, cpos = TRUE)
}
\arguments{
\item{s_gr}{Sample GenomicRanges object}

\item{ref_gr}{Reference GenomicRanges object}

\item{cpos}{cpos}
}
\value{
A GRanges intersected between s_gr and ref_gr with annotations
from ref_gr
}
\description{
Intersects the sample GR (s_gr) with the 
reference GR (ref_gr) and annotates all of the overlaps
in the sample GR with the reference
}
