% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFC.R
\name{mapSFC}
\alias{mapSFC}
\title{Map a space-filling curve using a HilbertCurve object}
\usage{
mapSFC(sfc = "sweep", hc_ord = NULL, order = NULL, uids = NULL)
}
\arguments{
\item{sfc}{space filling curve [character]}

\item{hc_ord}{dataframe object from setupRefHcMatrix(order=ord)$ord}

\item{order}{Order to run setupRefHcMatrix if hc_ord not given [integer]}

\item{uids}{Order to fill in the space-filling curve to avoid regenerating}
}
\value{
a dataframe object of a reordered X_Y positions correspondign to a different
space-filling curve for the same square matrix
}
\description{
Maps a space-filling curve to the HilbertCurve object
}
