% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvFiles.R
\name{mvFileToOncocode}
\alias{mvFileToOncocode}
\title{Customized code to move files to their oncocode directory}
\usage{
mvFileToOncocode(x, meta_df, datasets, pattern, sfc, cntype, rmfile = TRUE)
}
\arguments{
\item{x}{sample id}

\item{meta_df}{metadata data frame}

\item{datasets}{vector of datasets to use}

\item{pattern}{Pattern to gsub file IDs}

\item{sfc}{SFC type (sweep or hilbert, path related)}

\item{cntype}{CN type (TCN or ASCN, path related)}

\item{rmfile}{remove the file after copying or not}
}
\description{
Customized code to move files to their oncocode directory
}
