% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onco_meta}
\alias{onco_meta}
\title{Mapping between TCGA IDs, cell line IDs, filenames and cancer type}
\format{A list with 2 elements containing dataframes:
\describe{
  \item{ccl}{data frame of onco_meta_df}
  \item{tcga}{2 column dataframe of sampleID to oncocode}
}}
\source{
\url{}
}
\usage{
onco_meta
}
\description{
A list containing the onco_meta_df dataframe, or a TCGA metadata
dataframe containing the mapping between sample IDs and their
oncotree code.
}
\keyword{datasets}
