% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFCviz.R
\name{plotEuclidDist}
\alias{plotEuclidDist}
\title{Visualizes the euclidean distance between 2D space-filling curves}
\usage{
plotEuclidDist(xdist, frac_df)
}
\arguments{
\item{xdist}{distance matrix between a two-column dataframe}

\item{frac_df}{data frame to plot the chromosomes along the x and y-
axis with a corresponding colour}
}
\description{
Plots the euclidean distance matrix corresponding to 
the results from as.matrix(dist(data.frame(x1, y1))). 'frac_df' acts
as a supporting dataframe to label the chromosomes along the x and
y-axes.
}
