% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFC.R
\name{setupRefHcMatrix}
\alias{setupRefHcMatrix}
\title{Parse the HilbertCurve object}
\usage{
setupRefHcMatrix(order = 8)
}
\arguments{
\item{order}{Order to generate the HilbertCurve object [integer]}
}
\value{
A list object with 3 elements:
   - 'hc' = HilbertCurve object
   - 'mat' = Square matrix with unique X_Y coordinates (e.g. 1_2)
   - 'ord' = A data frame that maps to the 'mat' matrix. It dictates
   the genomic position for each unique X_Y coordinate
}
\description{
Wrapper function to run the HilbertCurve() function
but also to parse the returning object into a dataframe that 
maps the coordinates to genomic positions
}
