% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfcCurve.R
\name{SfcCurve}
\alias{SfcCurve}
\title{A copy of HilbertCurve() that lets POS inputs}
\usage{
SfcCurve(
  s,
  e,
  level = 4,
  mode = c("normal", "pixel"),
  reference = FALSE,
  reference_gp = gpar(lty = 3, col = "#999999"),
  arrow = TRUE,
  zoom = NULL,
  newpage = TRUE,
  background_col = "transparent",
  background_border = NA,
  title = NULL,
  title_gp = gpar(fontsize = 16),
  start_from = c("bottomleft", "topleft", "bottomright", "topright"),
  first_seg = c("horizontal", "vertical"),
  legend = list(),
  padding = unit(2, "mm"),
  sfc_pos = NULL
)
}
\arguments{
\item{s}{position that will be mapped as the start of the Hilbert curve. The value should a single numeric value.
If it is a vector, the minimum is used.}

\item{e}{position that will be mapped as the end of the Hilbert curve. The value should a single numeric value.
If it is a vector, the maximum is used.}

\item{level}{iteration level of the Hilbert curve. There will by ``4^level - 1`` segments in the curve.}

\item{mode}{"normal" mode is used for low ``level`` value and "pixel" mode is always used for high ``level`` value,
so the "normal" mode is always for low-resolution visualization while "pixel" mode is used for high-resolution visualization.
See 'details' for explanation.}

\item{reference}{whether add reference lines on the plot. Only works under 'normal' mode. The reference line
is only used for illustrating how the curve folds.}

\item{reference_gp}{graphic settings for the reference lines. It should be specified by `grid::gpar`.}

\item{arrow}{whether add arrows on the reference line. Only works under 'normal' mode.}

\item{zoom}{Internally, position are stored as integer values. To better map the data to the Hilbert curve, 
the original positions are zoomed
according to the range and the level of Hilbert curve. E.g. if 
the curve visualizes data ranging from 1 to 2 but level of the curve is set to 4,
the positions will be zoomed by ~x2000 so that values like 1.5, 1.555 can be mapped
to the curve with more accuracy. You don't need to care the zooming thing, 
proper zooming factor is calculated automatically.}

\item{newpage}{whether call `grid::grid.newpage` to draw on a new graphic device.}

\item{background_col}{background color.}

\item{background_border}{background border border.}

\item{title}{title of the plot.}

\item{title_gp}{graphic parameters for the title. It should be specified by `grid::gpar`.}

\item{start_from}{which corner on the plot should the curve starts?}

\item{first_seg}{the orientation of the first segment.}

\item{legend}{a `grid::grob` object, a `ComplexHeatmap::Legends-class` object, or a list of them.}

\item{padding}{padding around the Hilbert curve.}

\item{sfc_pos}{a dataframe object of x1, x2, y1, y2 positions according to POS}
}
\description{
A copy of HilbertCurve() that lets POS inputs
}
