% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFC.R
\name{splSquareFromBins}
\alias{splSquareFromBins}
\title{Split bins based on square overlap}
\usage{
splSquareFromBins(xleft, xright, ytop, ybottom, bins)
}
\arguments{
\item{xleft}{integer position of xleft}

\item{xright}{integer position of}

\item{ytop}{integer position of}

\item{ybottom}{integer position of}

\item{bins}{ordered dataframe from setupRefHcMatrix(order=ord)}
}
\value{
2 element list
  `TRUE`:  section of the dataframe that overlaps the input square
  `FALSE`: section of the dataframe that does not overlap the input square
}
\description{
Return the coordinates of a square that 
overlaps the 2d representation
}
