% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFC.R
\name{genHC}
\alias{genHC}
\title{Initiailize HilbertCurve object}
\usage{
genHC(end, order, mode = "pixel", sfc_pos = NULL)
}
\arguments{
\item{end}{Typically, the size of the genome [integer]}

\item{order}{Order to generate hilbert curve for [integer]}

\item{mode}{'pixel' or 'normal', default is 'pixel'}

\item{sfc_pos}{POS values for SFC mapping [dataframe]}
}
\value{
an object from HilbertCurve::HilbertCurve()
}
\description{
Wrapper function to run the HilbertCurve() function
using only a set end point and order
}
